app/requirements.txt
fastapi
uvicorn[standard]
azure-storage-blob
azure-search-documents
openai
pypdf
gunicorn
PyMuPDF